OD-17524 Move IshTaskChange.java and NumberedFilesComparator.java to types module
