+++ server/src/main/java/ish/oncourse/server/upgrades/liquibase/change/AddDefaultProductForms.java
@@ -8,6 +8,7 @@
 
 package ish.oncourse.server.upgrades.liquibase.change;
 
+import ish.liquibase.IshTaskChange;
 import ish.oncourse.server.ICayenneService;
 import ish.oncourse.server.cayenne.*;
 import ish.oncourse.server.db.SchemaUpdateService;
