+++ server/src/main/java/ish/oncourse/server/upgrades/liquibase/change/ClusteredAutoincrementServiceInit.java
@@ -10,6 +10,7 @@
  */
 package ish.oncourse.server.upgrades.liquibase.change;
 
+import ish.liquibase.IshTaskChange;
 import liquibase.database.Database;
 import liquibase.database.jvm.JdbcConnection;
 import liquibase.exception.CustomChangeException;
