+++ server/src/main/java/ish/oncourse/server/upgrades/liquibase/change/UpdateCreatedOnModifiedOnColumns.java
@@ -8,6 +8,7 @@
 
 package ish.oncourse.server.upgrades.liquibase.change;
 
+import ish.liquibase.IshTaskChange;
 import liquibase.database.Database;
 import liquibase.database.jvm.JdbcConnection;
 import liquibase.exception.CustomChangeException;
